freertos: assert: Always hang when assert is false
Zephyr's implentation makes the application stops right away when an
assert() is not verified.

As far as FreeRTOS is concerned, the behavior depends whether the
application is built in _release_ or _debug_ mode (assert() only hangs
for the latter).

Let's harmonize this assert() call and always stop the application and
avoid any confusion.

Signed-off-by: Stephane Viau <stephane.viau@nxp.com>