freertos: Main functions should return a value
Even though main() functions don't explicitly require a return statement
(default to 0 in C99), it is good practice to do so, especially since we
have a variable that reflects the status of the functions being called.

As a side note, this change gets rid of a warning by using the xResult
variable, which is currently set but unused when the assert() macro is
not implemented - in release mode.

Signed-off-by: Stephane Viau <stephane.viau@nxp.com>