rt_latency: Let "IRQ load" test case be executed at runtime
Today, the Test Case load conditions (cpu load, irq load, etc...) are
set at compile time which is not practical when a single test app might
be releases at the end.

The idea is to embed all the code for each test case and select at
runtime which test case is being executed (either through IPC command
line or sequentially with a for loop).

Signed-off-by: Stephane Viau <stephane.viau@nxp.com>