freertos: Ensure sections to be copied are 64-bit aligned
Some sections are are embedded in the code to be copied over to their
final location (in DDR, OCRAM, DTCM, etc...). The current copy method
assumes that the start ad end pointers of these sections are 64-bit
aligned, which is not obviously set by the linker script.

This patch explicitly align the end pointer to 64-bit.

Signed-off-by: Stephane Viau <stephane.viau@nxp.com>