rt_latency: Move the execution control loop outside of common code
Let's move all the initialization stuff from rt_latency_test() and
friends to a single function that shall be used for this kind of stuff:
rt_latency_init(). The advantage is that init/destroy can be more or
less symmetrical and - more important - the while loop of each task can
be moved outside of this common code. That allows the caller to control
when to start & stop each task execution properly.

Signed-off-by: Stephane Viau <stephane.viau@nxp.com>