script: update inmate entry address
Inmate entry address is now SOC dependent.

Changes to harpoon_set_configuration.sh:
No longer a template that needs to be instantiated. This is now a script
that runs directly on target and detects on which SOC it runs.

Changes to default harpoon.conf:
Use a template flag instead of hardcoding the inmate entry address.

Signed-off-by: Fabrice Goucem <fabrice.goucem@nxp.com>