common: logging: Add sdk_printf() support
The gen_avb stack uses sdk_printf() and sdk_vprintf() in the file
freertos/log.c.
Those are currently defined in the original FreeRTOS apps repository,
which we do not use.
We hence need the same definition in harpoon-apps so that the gen_avb
component can link properly with those two functions.

Signed-off-by: Stephane Viau <stephane.viau@nxp.com>