industrial: avb_tsn: Increase size of the task name
CID 10249727 complains that this buffer is too small when actually used
by FreeRTOS, which tried and copy configMAX_TASK_NAME_LEN (20)
characters for the task name. The tools doesn't see that the FreeRTOS
implementation actually "breaks" as soon as a '0' character is found,
which should make this issue a false positive... but why not increasing
to 20 characters to mae the tool happy and the size slightly bigger.

Signed-off-by: Stephane Viau <stephane.viau@nxp.com>