audio: common: Synchronize AVTP source stream
In Harpoon, we don't have the "processing period" aligned with the batch size.
Also, we are in polling mode, reading fixed amount based on application
processing period.
At this moment, we don't have clock recovery so we will overflow/underflow
regularly and we don't have synchronization based on gptp/stream timestamp yet.

Hence, we need to temporarily trick the data path to make it wait a bit
when we are ahead of time.

Signed-off-by: Stephane Viau <stephane.viau@nxp.com>