audio_smp: fix pll setup corner case
Fix corner case in audio SMP case: pll_element is located in the second pipeline,
but the id setting of this element was only done in the first pipeline.
Fix consist in broadcasting the pll id setting function in all the pipelines and
let the ctrl function handle the case where the element is not found and remove
the workaround that did the sai hardware setup only in the first pipeline,
as we can do it multiple times with no error and that the second pipeline is
initialized only at the second call of sai_setup.

Signed-off-by: Aziz Sellami <aziz.sellami@nxp.com>