audio: common: sai_drv: SAI stricter disable
Disable TE flag and starts a Software Reset (SR) on the SAI.
Clear flags and keep configuration registers, to avoid
reconfiguration. All operations are done simultaneously as
this seems to lead to a better behaviour.
Removed the busyloop, that in some cases blocked the
application, and replaced with a Software Reset.

Signed-off-by: Aziz Sellami <aziz.sellami@nxp.com>