audio: zephyr: main: fix thread ID passed to audio processing loop
The thread ID passed to the audio processing loop was retrieved from a pointer
to a variable from the stack, that changed during the execution of the main
function, and the data threads were started afterwards, which could lead to
a random number passed to the audio processing loop (depending on the for loop
execution). We pass directly the ID as a parameter.

Signed-off-by: Aziz Sellami <aziz.sellami@nxp.com>