industrial: Import motor control demo source code
The source code was imported from the following repo/version:
repo: freertos_avb_apps.git
tag: 6_0_0
path: boards/src/demo_apps/avb_tsn/tsn_app
All files related to the io_device role on the motor_control application
have not been imported since the goal is to port the application in
controller role ONLY (for the moment, at least).
This decision was made based on the fact that the i.MX 93 EVK board is
not compatible with the FRDM-MC-LVPMSM board that drives power to the
motor.

Signed-off-by: Caio Molinari <caioheitor.barzmolinari@nxp.com>