industrial: Add support to motor control usecase
Modify and add files to make tsn_app_motor_control Harpoon compatible.
By default, the build will add the motor control application
The motor control is a particular use-case of the
more general avb_tsn use-case. To launch app mode MOTOR_NETWORK,
the user must now provide additional flags to define the app_mode,
the desired control strategy and the number of io_devices
(naturally, this last option is used on NETWORK_ONLY mode as well) of
the application: harpoon_ctrl ethernet -r 0 -m 0 -c 0 -n 1. If the -m
flag is not provided, the default app_mode is NETWORK_ONLY. For the moment,
the motor control application is only supported on Harpoon on the role of
CONTROLLER.

Signed-off-by: Caio Molinari <caioheitor.barzmolinari@nxp.com>