audio: sai: add support for channel masking
The CS42448 on the MX93AUD-HAT only supports TDM with 8 slots (and 32 bits
width), but the audio board has only 6 input channels (and 8 output
channels).
So, let's leverage the channel masking feature of the SAI to avoid using
extra useless buffers and have only valid/useful data in SAI FIFOs (and
thus configure the watermark according to effective channel count).

Signed-off-by: Aziz Sellami <aziz.sellami@nxp.com>