industrial: fix wrong pointer for event message queue
The bad pointer usage is causing that CAN application on FreeRTOS is showing
empty stats (no irq, no errors ...) and it's crashing on Zephyr with an invalid
spinlock assertion. The issue originates from the move to use
rtos-abstraction-layer with the dynamic message queue instead of the old static
one.  It's affecting CAN only since the event_data() handling is only used for
that use case.

fixes: (1355c502c7e audio/industrial: replace os_mqueue by
rtos-abstraction-layer)

Signed-off-by: Aziz Sellami <aziz.sellami@nxp.com>