industrial: fix mode index check
Harpoon cell crashes when running the command on Linux:
harpoon_ctrl can -r 1 -n 0
The check  on the mode index is done with ARRAY_SIZE, which takes the
biggest array size on the struct, which here is the ethernet use case
and create a mismatch of the array size.
Introducing an explicit array size to prevent this issue.

Signed-off-by: Aziz Sellami <aziz.sellami@nxp.com>