industrial: can: zephyr: imx95lpd5evk19 add support
Add support for the CAN application on harpoon-apps for zephyr.

Use the CAN2 interface for the following reasons:
 - CAN2 is located on the WAKEUPMIX, reducing overlap with the AONMIX,
   in which the CM33 is located.
 - Linux use case is using the same interface, so it will be easier to
   cross-check.

Configure the CAN hardware through Zephyr. Using the hal_nxp as for
other implementation (evkmim8mp and mcimx93evk), would have introduced
dependencies on hal_nxp mcux-sdk drivers for the System Manager that
are currently not provided.  Using the Zephyr drivers for this task is
more straight-forward as it would need less hal_nxp drivers and will
not introduce overlap in the System Manager usage (mailbox, irq, ...).
A new zephyr driver binding is introduced for the fake
"nxp,imx-flexcan" driver, so that pinctrl and clock devices are
declared and are accessible through device tree APIs.

A new abstraction for Flexcan core clock rate access is introduced, as
we cannot use the hal_nxp API for imx95lpd5evk19, due to missing System
Manager drivers.

Signed-off-by: Aziz Sellami <aziz.sellami@nxp.com>