industrial: imx95lp4xevk15: freertos: enable ethernet TSN application
Add support on i.MX 95 15x15 LPDDR4x EVK for FreeRTOS on harpoon-apps.

Support the ENET1 interface (J8) of the i.MX 95 15x15 LPDDR4x EVK,
which uses the ENETC0 1G interface.
This implementation relies on the NETC IP and uses directly the ENETC0
PSI. For the 1588 Timers used in the genAVB application, we use the
NETC 1588 Timer and rely on the MSGINTR IP to generated the needed
interrupts for event handling.
Use the Port EMDIO peripheral for proper EMDIO management and ethernet
interface configuration.

Signed-off-by: Aziz Sellami <aziz.sellami@nxp.com>