industrial: imx95lp4xevk15: zephyr: enable ethernet TSN application
Add support on i.MX 95 15x15 LPDDR4x EVK for Zephyr on harpoon-apps.
This implementation uses the same interfaces as the FreeRTOS
implementation: ENETC0 interface (J8) and NETC 1588 timer.

Adapt the code for pinmux and clock handling and use OS-specific
abstraction so that Zephyr can rely on its drivers for SCMI usage and
keep its ownership of this interface.

Add device tree NETC-specific nodes in overlay for MMU mapping, as well
as clock and pinctrl properties to initialize devices on application
side for devices that relies on hal_nxp drivers.

Introduce a unified binding file, to be shared across nodes that needs
clock and pinctrl properties, for devices that relies on hal_nxp
drivers.

Signed-off-by: Aziz Sellami <aziz.sellami@nxp.com>