audio: avtp_source: increase playback offset to cover for delay caused by batch misalignment between stack and application
The pipeline is driven by SAI interrupts based on the configured period
and the avtp souce element are reading audio samples from the GenAVB/TSN
stack by that match. OTOH, the stack is having its own packet batching
internally. So, in some cases where the application batch is close to the
stack batch (and eventually > stack_batch / 2) the application may encounter
an underflow on requesting data too early from the stack.
To mitigate that risk, increase the playback offset to cover an additional
500us (worst case period is 32 samples @ 44.1 Khz but 500 is sufficient for now)
The downside of this change is eventually an increase of end-to-end latency.
But, a proper fix should be provided later to align both the stack and application
on the processing batch and get rid of this fixed delay.

Signed-off-by: Marouen Ghodhbane <marouen.ghodhbane@nxp.com>