common: avb: enable wait for avdecc start command in AVB non-Milan mode
Since audio_avb_init() in rtos-apps always send the GENAVB_MSG_MEDIA_STACK_ENTITY_START
message after opening the media stack channel, enable the flag for non-Milan mode
as well.
Moreover, without any waitmask the entity would be started in AVDECC stack (on GM
status change for example) without even waiting on the opening of the media stack
channel and the IPC for setting the default clock source for the AEM would not be
delivered.
This change ensures that the application receives and handles the SET_CLOCK_SOURCE
message from the stack for the default clock source of the AEM after opening the
channel and issuing the AVDECC start command.

Signed-off-by: Marouen Ghodhbane <marouen.ghodhbane@nxp.com>